+++ src/HDInsight/HDInsight.Test/ScenarioTests/HDInsightClusterTests.ps1
@@ -275,3 +275,41 @@ function Test-CreateClusterWithKafkaRestProxy{
 		Remove-AzResourceGroup -ResourceGroupName $cluster.ResourceGroup
 	}
 }
+
+<#
+.SYNOPSIS
+Test Create Azure HDInsight Cluster with Relay Outbound and Private Link
+#>
+
+function Test-CreateClusterWithRelayOutoundAndPrivateLink{
+
+	# Create some resources that will be used throughout test
+	try
+	{
+		# prepare parameter for creating parameter
+		$params= Prepare-ClusterCreateParameter -location "South Central US"
+
+		# Private Link requires vnet has firewall, this is difficult to create dynamically, just hardcode here
+		$vnetId="/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/rg/providers/Microsoft.Network/virtualNetworks/fakevnet"
+		$subnetName="default"
+
+		# create cluster
+		$cluster = New-AzHDInsightCluster -Location $params.location -ResourceGroupName $params.resourceGroupName `
+		-ClusterName $params.clusterName -ClusterSizeInNodes $params.clusterSizeInNodes -ClusterType $params.clusterType `
+		-StorageAccountResourceId $params.storageAccountResourceId -StorageAccountKey $params.storageAccountKey `
+		-HttpCredential $params.httpCredential -SshCredential $params.sshCredential `
+		-MinSupportedTlsVersion $params.minSupportedTlsVersion `
+		-VirtualNetworkId $vnetId -SubnetName $subnetName -Version 3.6 `
+		-ResourceProviderConnection Outbound -PrivateLink Enabled
+
+		Assert-AreEqual $cluster.NetworkProperties.ResourceProviderConnection Outbound
+		Assert-AreEqual $cluster.NetworkProperties.PrivateLink Enabled
+		
+	}
+	finally
+	{
+		# Delete cluster and resource group
+		Remove-AzHDInsightCluster -ClusterName $cluster.Name
+		Remove-AzResourceGroup -ResourceGroupName $cluster.ResourceGroup
+	}
+}